+++ Runtime/Common/IComponent.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+
+namespace SimpleMan.Utilities
+{
+    public interface IComponent
+    {
+        //------PROPERTIES
+        GameObject SelfGameObject { get; }
+        bool PrintLogs { get; set; }
+    }
+}